pyproject.toml
[project]
name = "goodmoringtoyou"
version = "0.1.0"
description = ""
authors = [
    {name = "jayden",email = "jaydenhyunjinlee@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "boto3 (>=1.36.2,<2.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
